requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
openpyxl==3.1.2
colorama==0.4.6